requirements.txt
asgiref==3.5.2
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
defusedxml==0.7.1
Django==3.2.13
django-allauth==0.51.0
django-crispy-forms==1.14.0
django-markdownx==3.0.1
idna==3.3
importlib-metadata==4.11.4
Markdown==3.3.7
oauthlib==3.2.0
Pillow==9.1.1
pycparser==2.21
PyJWT==2.4.0
python3-openid==3.2.0
pytz==2022.1
requests==2.28.0
requests-oauthlib==1.3.1
sqlparse==0.4.2
typing-extensions==4.2.0
urllib3==1.26.9
waitress==2.1.2
whitenoise==6.2.0
zipp==3.8.0
